feat: update function navbar and remove usenavbar hook
